Add CI for the tests, and fix what it found
There was no workflow running `cargo test`. The only one built the ISO,
so a change that broke 531 tests would have gone green as long as the
image still booted.

The new `ci` workflow runs on every change, not only the ones that touch
the image:

  - build, test and clippy on the workspace, with warnings denied
  - `cargo check` for x86_64, arm64 and musl, because the kernel-facing
    backends are Linux-only and the ISO links a static musl binary
  - shellcheck over iso/, with each script checked as the shell its own
    shebang names

Turning warnings into errors found one: `libc::time_t` and
`libc::suseconds_t` are deprecated on musl, which is exactly the target
the ISO builds. The kernel writes a `struct timeval` into `input_event`
and its fields are kernel longs, so that is what they are now; musl's
own definitions changed width, the kernel's did not.

Also cleared the clippy warnings that were already there. `cargo fmt`
is deliberately not a gate yet: the tree is not rustfmt-clean, and a
repo-wide reformat belongs in its own change rather than buried in this
one.

+++ compositor/tos-render/src/surface.rs
@@ -248,9 +248,9 @@ impl<'a> Surface<'a> {
 /// Blend `color` over a packed XRGB pixel.
 #[inline]
 fn blend_packed(dest: u32, color: Rgb, alpha: u8) -> u32 {
-    let dr = ((dest >> 16) & 0xff) as u32;
-    let dg = ((dest >> 8) & 0xff) as u32;
-    let db = (dest & 0xff) as u32;
+    let dr = (dest >> 16) & 0xff;
+    let dg = (dest >> 8) & 0xff;
+    let db = dest & 0xff;
     let a = alpha as u32;
     let inv = 255 - a;
     let r = (color.r as u32 * a + dr * inv) / 255;
